[GREEN] Implement integration with auth for crud product

+++ src/main/java/id/ac/ui/cs/advprog/hoomgroombeadmin/controller/ProductRestController.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import id.ac.ui.cs.advprog.hoomgroombeadmin.model.Product;
 import id.ac.ui.cs.advprog.hoomgroombeadmin.service.ProductFilterContext;
 import id.ac.ui.cs.advprog.hoomgroombeadmin.service.ProductService;
-import lombok.RequiredArgsConstructor;
+import id.ac.ui.cs.advprog.hoomgroombeadmin.service.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.scheduling.annotation.Async;
@@ -24,22 +24,38 @@ public class ProductRestController {
 
     @Autowired
     private ProductService service;
+
+    @Autowired
+    private JwtService jwtService;
+
     private ProductFilterContext productFilterContext;
 
     @PostMapping("/create")
-    public ResponseEntity<Product> createProductPost(@RequestBody Product product) throws JsonProcessingException {
+    public ResponseEntity<Product> createProductPost(@RequestHeader (value = "Authorization") String token, @RequestBody Product product) throws JsonProcessingException {
+        token = token.substring(7);
+        if (!jwtService.isTokenValid(token) || jwtService.extractRole(token) != "ADMIN"){
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         Product result = service.save(product);
         return new ResponseEntity<>(result, HttpStatus.CREATED);
     }
 
     @PostMapping("/update/save")
-    public ResponseEntity<Product> updateProductPost(@RequestBody Product product) {
+    public ResponseEntity<Product> updateProductPost(@RequestHeader (value = "Authorization") String token, @RequestBody Product product) {
+        token = token.substring(7);
+        if (!jwtService.isTokenValid(token) || jwtService.extractRole(token) != "ADMIN"){
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         Product result = service.save(product);
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @GetMapping("/update/{productId}")
-    public ResponseEntity<Product> updateProductPage(@PathVariable String productId){
+    public ResponseEntity<Product> updateProductPage(@RequestHeader (value = "Authorization") String token, @PathVariable String productId){
+        token = token.substring(7);
+        if (!jwtService.isTokenValid(token) || jwtService.extractRole(token) != "ADMIN"){
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         Product result = service.getProductById(productId);
         if (result == null){
             return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
@@ -48,7 +64,11 @@ public class ProductRestController {
     }
 
     @DeleteMapping("/delete/{productId}")
-    public ResponseEntity<String> deleteProduct(@PathVariable String productId){
+    public ResponseEntity<String> deleteProduct(@RequestHeader (value = "Authorization") String token, @PathVariable String productId){
+        token = token.substring(7);
+        if (!jwtService.isTokenValid(token) || jwtService.extractRole(token) != "ADMIN"){
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
         String result = "Deleted product with ID " + service.delete(productId);
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
@@ -87,7 +107,12 @@ public class ProductRestController {
     @PostMapping("/sold")
     @Async
     public CompletableFuture<String> incrementSales(
+            @RequestHeader (value = "Authorization") String token,
             @RequestBody HashMap<String, Integer> productsSold){
+        token = token.substring(7);
+        if (!jwtService.isTokenValid(token)){
+            return CompletableFuture.failedFuture(new IllegalAccessException("Log in needed"));
+        }
         return service.incrementSales(productsSold);
     }
 }
